backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-Caching==2.1.0
python-dotenv==1.0.0
PyMySQL==1.1.0
numpy==1.24.3
pandas==2.0.3
pytest==7.4.3
pytest-flask==1.3.0
gunicorn==21.2.0
h5py==3.10.0
astropy==6.0.0